Allows non-root user to use Docker during run
When using the install-docker role, you're unable to consume Docker
commands from a non-root user during the same run. This change resets
the SSH connection to allow for the non-root user to have their groups
updated in the shell, allowing for execution of Docker commands during
the same run.

Only resets the connection when the docker group has been added to the
user during that run.

Adds a documentation WARNING to the top of the README file noting that
we setup the non-root user in the Docker group, which can have
security implications in some environments.

Closes #6