Remove superfluous ssh connection reset
Remove the usage of an ssh connection reset in order to load a group
into the running user. First of all, the way the reset was implemented
was causing issues with hung systems, requiring a hard reboot. Clearly
that isn't ideal. Secondly, the reset is was totally superfluous as
we can add the group to the currently running user with a 'newgrp'
command.